A lazy sample object

Often you will resample a dataset many times, as in cross-validation or bootstrapping. Copying and storing each complete sample is costly in both memory and time. This object instead stores the original data as a quosure and indexes, and only computes the sample object when needed.

Usage

smpl <- LazySample$new(data, idx, fun = NULL)
    smpl$idx
smpl$data
smpl$extractor
smpl$draw_sample()
smpl$sample
smpl$data
    print(smpl)
as.data.frame(smpl)
    

Arguments

data

An unevaluated expression that will be captured as a quosure.

idx

A character or integer vector. If not these, it will be coerced to an integer (numeric) or character vector (all other classes). Since data is not evaluated when the object is created, the validity of these indexes is not checked.

extractor

A function with two arguments (data, index) which is used to extract the sample when it is computed. This is coerced to a function via as_function, meaning that a formula like ~ .x[.y] can be used. If NULL, then extract_sample is used.

Details

$idx returns the indexes in the sample. $quo returns a quosure that returns the data to be sampled from when evaluated with eval_tidy. $data returns the data to be sampled from. $sample or $draw_sample() or draw_sample(smpl) returns the sample. $sample is is active binding and thus the sample is lazily calculated. print(smpl) or smpl$print() shows some information about the LazySample object on the screen. as.data.frame(smpl) returns the sample and coerces it to a data frame. This is a convenience function since many modeling functions, e.g. lm, call as.data.frame on the data argument.

See also

lazy_sample is a convenience function to create lazy sample objects, lazy_sample_lst creates a list of lazy samples objects, and lazy_sample_idx creates a data frame with a list column of lazy sample objects.

Examples

# New lazy sample object
smpl <- LazySample$new(~ mtcars, 1:5)
# Get the indexes
smpl$idx

#> [1] 1 2 3 4 5

# Get the expression containing the data
smpl$quo

#> <quosure: local>
#> ~mtcars

# the expression is stored as a quosure
class(smpl$quo)

#> [1] "quosure" "formula"

# with an expression
rlang::get_expr(smpl$quo)

#> mtcars

# and an environment
rlang::get_env(smpl$quo)

#> <environment: 0x115ff65b8>

# Get the sample - it is not computed until called
smpl$sample

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2

smpl$draw_sample()

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2

# Get the original data
smpl$data

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

# Return the sample as a data frame
as.data.frame(smpl)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2

# LazySample objects also work with non-data frames
x <- 1:100
smpl_x <- LazySample$new(~ x, sample(length(x), 10))
smpl_x

#> <lazy_sample>
#> data:  ~x 
#> idx [10]: 78, 42, 67, 48, 95, 62, 10, 1, 54, 23

smpl_x$sample

#>  [1] 78 42 67 48 95 62 10  1 54 23

y <- matrix(1:300, nrow = 100)
smpl_y <- LazySample$new(~ y, sample(nrow(y), 10))
smpl_y

#> <lazy_sample>
#> data:  ~y 
#> idx [10]: 90, 79, 2, 73, 83, 43, 52, 35, 68, 80

smpl_y$sample

#>       [,1] [,2] [,3]
#>  [1,]   90  190  290
#>  [2,]   79  179  279
#>  [3,]    2  102  202
#>  [4,]   73  173  273
#>  [5,]   83  183  283
#>  [6,]   43  143  243
#>  [7,]   52  152  252
#>  [8,]   35  135  235
#>  [9,]   68  168  268
#> [10,]   80  180  280

# the extractor function can be customized
# for example, randomly sample the columns of a data frame
z <- as.data.frame(matrix(1:1000, nrow = 10))
smpl_z <- LazySample$new(~ z, sample(nrow(y), 10),
                         extractor = ~ .x[ , .y, drop = FALSE])
smpl_z

#> <lazy_sample>
#> data:  ~z 
#> idx [10]: 80, 19, 49, 82, 92, 43, 37, 7, 34, 47

smpl_z$sample

#>    V80 V19 V49 V82 V92 V43 V37 V7 V34 V47
#> 1  791 181 481 811 911 421 361 61 331 461
#> 2  792 182 482 812 912 422 362 62 332 462
#> 3  793 183 483 813 913 423 363 63 333 463
#> 4  794 184 484 814 914 424 364 64 334 464
#> 5  795 185 485 815 915 425 365 65 335 465
#> 6  796 186 486 816 916 426 366 66 336 466
#> 7  797 187 487 817 917 427 367 67 337 467
#> 8  798 188 488 818 918 428 368 68 338 468
#> 9  799 189 489 819 919 429 369 69 339 469
#> 10 800 190 490 820 920 430 370 70 340 470