Generate cross-validation test-training sets

Generate cross-validation test/training pairs by specifying the indexes in the elements in the test and training sets. This function provides a manual method for generating data frames with with cross-validation resample objects, while other functions, such as crossv_kfold, algorithmically handle generate the indexes in the test- and training sets.

crossv_idx(expr, train, test, extractor = NULL)

Arguments

expr
A quosure

train
A list of integer or character vectors

test
A list of integer or character vector

extractor
A function or object coercible to a function by as_function. See LazySample for more information.

Value

A data frame with <

train

A list column with objects representing the training sets. The list elements are lazy_sample objects.

test

A list column with objects representing the training sets. The list elements are lazy_sample objects.

References

Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.

Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.

Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.

See also

Other cross-validation functions: crossv_lpo, crossv_mc

Examples

# specify test and training partitions
test <- list(1:10, 10:20, 30:nrow(mtcars))
train <- purrr::map(test, setdiff, y = test)
crossv_idx(mtcars, train = train, test = test)

#> # A tibble: 3 x 2
#>           train          test
#>          <list>        <list>
#> 1 <lazy_sample> <lazy_sample>
#> 2 <lazy_sample> <lazy_sample>
#> 3 <lazy_sample> <lazy_sample>