Generate cross-validated K-fold test-training pairs

Generate cross-validated K-fold test-training pairs. In addition to ordinary K-fold cross-validation, this supports stratified K-fold cross validation if data is a grouped data frame and stratify = TRUE, and Group K-fold if data is a grouped data frame and stratify = FALSE.

crossv_kfold(expr, K, ...)

crossv_kfold_n(n, K = 5L, shuffle = TRUE)

Arguments

expr
An expression for the data to be sampled. This does not need to be a data frame, as lazy_sample objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a quosure.

K
An integer scalar. The number of cross-validation folds.

...
Arguments passed to methods

n
A non-negative scalar integer representing the number of elements to sample from. If n is NULL, then expr is evaluated and called be idx_len to calculate this number. However, if n is non-NULL, then expr is not evaluated at all until samples are requested.

shuffle
A logical scalar. If TRUE, then the elements are randomly shuffled prior to partitioning into folds.

Value

A data frame (tibble) with K rows and the following columns:

train

A list column with objects representing the training sets. For crossv_kfold, the elements are vectors of the indexes; for crossv_kfold_n, the elements are lazy_sample objects.

test

A list column with objects representing the training sets. For crossv_kfold, the elements are vectors of the indexes; for crossv_kfold_n, the elements are lazy_sample objects.

References

Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.

Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.

Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.

See also

The boot function cv.glm and modelr function crossv_kfold for other implementations of k-fold cross-validation.