Generate K cross-validation test-training pairs

crossv_mc generates random train/test pairs where either proportion frac of the elements are in the test set and 1 - frac elements are in the training set, or size elements are in the test set and n - size elements are in the training set.

crossv_mc(expr, ..., n = NULL, extractor = NULL)

crossv_mc_n(n, K = 10L, frac = 0.3, size = NULL, prob = NULL)

Arguments

expr
An expression for the data to be sampled. This does not need to be a data frame, as lazy_sample objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a quosure.

...
Arguments passed to crossv_mc_n.

n
A non-negative scalar integer representing the number of elements to sample from. If n is NULL, then expr is evaluated and called be idx_len to calculate this number. However, if n is non-NULL, then expr is not evaluated at all until samples are requested.

extractor
A function or object coercible to a function by as_function. See LazySample for more information.

K
Number of test/train splits to generate.

frac
A numeric scalar between 0 and 1. The fraction of elements in the test set.

size
An integer scalar. The number of elements in the test set. If size is non-NULL, then it takes preference over frac.

prob
Probability weight that an element is in the test set.

Value

A data frame (tibble) with K rows and the following columns:

train

A list column with objects representing the training sets. For crossv_mc, the elements are vectors of the indexes; for crossv_mc_n, the elements are lazy_sample objects.

test

A list column with objects representing the training sets. For crossv_mc, the elements are vectors of the indexes; for crossv_mc_n, the elements are lazy_sample objects.

References

Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.

Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.

Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.

See also

The modelr function crossv_mc.

Other cross-validation functions: crossv_idx, crossv_lpo

Examples

library("purrr")
library("dplyr")

# 5-fold cross-validation
cv1 <- crossv_mc(mtcars)
models <- map(cv1$train, ~ lm(mpg ~ wt, data = .$sample))
summary(map2_dbl(models, map(cv1$test, ~ .$sample), modelr::rmse))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   2.475   2.595   2.799   3.103   3.597   4.058 

# crossv_kfold_n works with a number of elements
cv2 <- crossv_mc_n(100)
cv2

#> # A tibble: 10 x 2
#>         train      test
#>        <list>    <list>
#>  1 <int [94]> <int [6]>
#>  2 <int [94]> <int [6]>
#>  3 <int [94]> <int [6]>
#>  4 <int [94]> <int [6]>
#>  5 <int [94]> <int [6]>
#>  6 <int [94]> <int [6]>
#>  7 <int [94]> <int [6]>
#>  8 <int [94]> <int [6]>
#>  9 <int [94]> <int [6]>
#> 10 <int [94]> <int [6]>

cv2$train[[1]]

#>  [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  19  20
#> [20]  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#> [39]  40  41  42  43  44  46  47  48  49  50  51  52  53  54  55  56  57  58  59
#> [58]  60  61  62  63  64  65  66  67  68  70  71  72  73  74  75  76  77  80  81
#> [77]  82  83  84  85  86  87  88  90  91  92  93  94  95  96  97  98  99 100

cv2$test[[2]]

#> [1] 13 54 29 78 14 83