Extract samples from objects

The generic function extract_sample is the function that extracts the elements specified by idx from the the data in a lazy_sample object.

extract_sample(x, idx, ...)

# S3 method for default
extract_sample(x, idx, margin = 1L, ...)

# S3 method for matrix
extract_sample(x, idx, rows = TRUE, drop = FALSE, ...)

# S3 method for data.frame
extract_sample(x, idx, rows = TRUE, drop = FALSE, ...)

# S3 method for quosure
extract_sample(x, ...)

Arguments

x
An object

idx
An integer or character vector representing index values

...
Arguments used by methods.

margin
An integerish scalar. The margin (dimension) to which idx refer. For example, extract(x, idx, 1) is like x[idx, ], extract(x, idx, 2) is like x[ , idx, ...], and so on.

rows
A logical scalar. If TRUE, then extract rows, else extract columns.

drop
A logical scalar. If TRUE, then coerce to the lowest possible dimension. See drop.

Details

It is largely for internal use in order to abstract the details for extracting elements from data frames, vectors, matrices, arrays, etc. However, it is exported so that the user can add methods for new objects or change the default behavior. The extractor field in lazy_sample objects and the extractor argument in many functions will override this function.

Methods (by class)

default: The default method is equivalent to x[idx] when there are no dimensions in x. When there are dimensions in x, then idx refers to the margin dimension.

matrix: The default method for matrices extracts rows by default and does not drop dimensions.

data.frame: The data frame method extracts rows by default, and does not drop dimensions.

quosure: For quosures, the method will evaluate the expression and then call the next method on the result.

Examples

require("tibble")
require("rlang")

#> Loading required package: rlang

#> 
#> Attaching package: 'rlang'

#> The following object is masked from 'package:lazysample':
#> 
#>     is_named

#> The following object is masked from 'package:tibble':
#> 
#>     has_name

#> The following objects are masked from 'package:purrr':
#> 
#>     %@%, %||%, as_function, flatten, flatten_chr, flatten_dbl,
#>     flatten_int, flatten_lgl, invoke, is_atomic, is_bare_atomic,
#>     is_bare_character, is_bare_double, is_bare_integer, is_bare_list,
#>     is_bare_logical, is_bare_numeric, is_bare_vector, is_character,
#>     is_double, is_empty, is_formula, is_function, is_integer, is_list,
#>     is_logical, is_null, is_scalar_atomic, is_scalar_character,
#>     is_scalar_double, is_scalar_integer, is_scalar_list,
#>     is_scalar_logical, is_scalar_vector, is_vector, list_along,
#>     prepend, rep_along, set_names, splice

#> The following objects are masked from 'package:testthat':
#> 
#>     is_false, is_null, is_true

#> The following objects are masked from 'package:devtools':
#> 
#>     ns_env, pkg_env

# Extract indexes from vectors
extract_sample(1:100, 1:10)

#>  [1]  1  2  3  4  5  6  7  8  9 10

# For arrays: by default, extract from the first dimension
extract_sample(array(seq_len(60), dim = c(5, 4, 3)), 1:3)

#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    6   11   16
#> [2,]    2    7   12   17
#> [3,]    3    8   13   18
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   21   26   31   36
#> [2,]   22   27   32   37
#> [3,]   23   28   33   38
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   41   46   51   56
#> [2,]   42   47   52   57
#> [3,]   43   48   53   58
#> 

# For matrices: by default, extract rows
extract_sample(array(seq_len(20), dim = c(10, 2)), 1:5)

#>      [,1] [,2]
#> [1,]    1   11
#> [2,]    2   12
#> [3,]    3   13
#> [4,]    4   14
#> [5,]    5   15

# For data frames: by default, extract rows
extract_sample(tibble(foo = 1:10, bar = 1:10), 1:5)

#> # A tibble: 5 x 2
#>     foo   bar
#>   <int> <int>
#> 1     1     1
#> 2     2     2
#> 3     3     3
#> 4     4     4
#> 5     5     5

# For quosures: by default, evaluate first
x <- tibble(foo = 1:10, bar = 1:10)
q <- quo(x)
q

#> <quosure: local>
#> ~x

extract_sample(q, 1:5)

#> # A tibble: 5 x 2
#>     foo   bar
#>   <int> <int>
#> 1     1     1
#> 2     2     2
#> 3     3     3
#> 4     4     4
#> 5     5     5