Generate the number of elements to sample from

This method provides number of elements from which to sample. This number is needed by the resampling functions. This is better than explicit enumeration of indexes, since resampling does not require the explicit enumeration.

idx_len(x, ...)

# S3 method for default
idx_len(x, margin = 1L, ...)

# S3 method for array
idx_len(x, margin = 1L, ...)

# S3 method for matrix
idx_len(x, rows = TRUE, ...)

# S3 method for data.frame
idx_len(x, rows = TRUE, ...)

# S3 method for quosure
idx_len(x, ...)

Arguments

x
An object

...
Arguments passed to other methods

margin
Integer scalar; the array margin to use

rows
Logical scalar. If TRUE, use return number of rows, else the number of columns.

Value

A scalar integer

Methods (by class)

default: The default is to use the length of the object.

array: For arrays, specify the margin for the number of elements. The default is the first margin.

matrix: For matrices, there are two dimensions, so specify whether the number of rows or columns are to be used.

data.frame: For data frames, there are two dimensions, so specify whether the number of rows or columns are to be used.

quosure: For quosures, the expression is evaluated and then idx_len is called on the result.

Examples

require("tibble")
require("rlang")

# Extract number of indexes for various classes

# default (vectors)
idx_len(seq_len(10))

#> [1] 10

# matrices (rows by default)
idx_len(matrix(seq_len(20), nrow = 2))

#> [1] 2

# arrays (first dimension by default)
idx_len(array(seq_len(60), dim = c(5, 4, 3)))

#> [1] 5

# data frames (rows by default)
idx_len(tibble(foo = 1:10, bar = 1:10))

#> [1] 10

# quosures are evaluated first
x <- tibble(foo = 1:10, bar = 1:10)
q <- quo(x)
q

#> <quosure: local>
#> ~x

idx_len(q)

#> [1] 10