Create lazy sample objects

The function lazy_sample is a convenience function to create LazySample objects. The function lazy_sample_lst creates a list of lazy sample objects with the same expression, but multiple indexes. The function lazy_sample_idx creates a data frame with a list-column of LazySample objects.

lazy_sample(expr, idx, extractor = NULL)

lazy_sample_lst(expr, idx, extractor = NULL)

lazy_sample_idx(expr, idx, extractor = NULL)

Arguments

expr
An expression with the data.

idx
A list of integer or character vector with the indexes in each sample.

extractor
A function or object coercible to a function by as_function. See LazySample for more information.

Value

lazy_sample

A LazySample object.

lazy_sample_lst

A list of LazySample objects with length equal to the length of idx.

lazy_sample_idx

A data frame with one column and rows equal to the length of idx. The single column is named sample and is a list of LazySample objects.

See also

LazySample for detailed information on the lazy_sample class.

Examples

require("purrr")

# lazy_sample is a convenience function to create lazy_sample objects
lazy_sample(mtcars, 1:5)

#> <lazy_sample>
#> data:  ~mtcars 
#> idx [5]: 1, 2, 3, 4, 5

# lazy_sample_lst creates a list of lazy_sample objects
lazy_sample_idx(mtcars, rerun(3, nrow(mtcars)))

#> # A tibble: 3 x 1
#>          sample
#>          <list>
#> 1 <lazy_sample>
#> 2 <lazy_sample>
#> 3 <lazy_sample>

# lazy_sample_idx creates a data frame with a list column
# of lazy_sample objects
lazy_sample_idx(mtcars, rerun(3, nrow(mtcars)))

#> # A tibble: 3 x 1
#>          sample
#>          <list>
#> 1 <lazy_sample>
#> 2 <lazy_sample>
#> 3 <lazy_sample>