Generate permutation subsets

Generate R replicates with permutations of the rows or groups of a data frame.

permute(expr, R = 1L, n = NULL, extractor = NULL)

permute_n(n, R = 1L)

Arguments

expr
An expression for the data to be sampled. This does not need to be a data frame, as lazy_sample objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a quosure.

R
The number of permutations to generate.

n
A non-negative scalar integer representing the number of elements to sample from. If n is NULL, then expr is evaluated and called be idx_len to calculate this number. However, if n is non-NULL, then expr is not evaluated at all until samples are requested.

extractor
A function or object coercible to a function by as_function. See LazySample for more information.

...
Arguments passed to methods

Value

data frame (tibble) with R rows and the following columns.

sample

A list column with objects representing the training sets. For permute, the elements are vectors of the indexes; for permute_n, the elements are lazy_sample objects.